Move keyword handler to ui.go
