+++ cmd/dgraph/main.go
@@ -658,48 +658,6 @@ func (s *grpcServer) Run(ctx context.Context,
 	return resp, err
 }
 
-type keyword struct {
-	// Type could be a predicate, function etc.
-	Type string `json:"type"`
-	Name string `json:"name"`
-}
-
-type keywords struct {
-	Keywords []keyword `json:"keywords"`
-}
-
-// Used to return a list of keywords, so that UI can show them for autocompletion.
-func keywordHandler(w http.ResponseWriter, r *http.Request) {
-	addCorsHeaders(w)
-	// TODO: Remove this code and replace with query from ui
-	preds := schema.State().Predicates(1)
-	kw := make([]keyword, 0, len(preds))
-	for _, p := range preds {
-		kw = append(kw, keyword{
-			Type: "predicate",
-			Name: p,
-		})
-	}
-	kws := keywords{Keywords: kw}
-
-	predefined := []string{"id", "_uid_", "after", "first", "offset", "count",
-		"@facets", "@filter", "func", "anyofterms", "allofterms", "anyoftext", "alloftext", "leq", "geq", "or", "and",
-		"orderasc", "orderdesc", "near", "within", "contains", "intersects"}
-
-	for _, w := range predefined {
-		kws.Keywords = append(kws.Keywords, keyword{
-			Name: w,
-		})
-	}
-	js, err := json.Marshal(kws)
-	if err != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		w.Write([]byte(err.Error()))
-		return
-	}
-	w.Write(js)
-}
-
 func checkFlagsAndInitDirs() {
 	if len(*cpuprofile) > 0 {
 		f, err := os.Create(*cpuprofile)
